First JUnit Tests
Started on tests for Game. Realized it'd be a lot of them, so I split
them into separate test classes categorically by methods (Game based,
Player-Specific, and Character-Specific). Note: I did not write all of
them yet.

Also started on tests for Constructs.

With both, I realized that since we're actively changing the structure
of classes, it's difficult to write the tests due to method names and
stuff changing.

I've typed up a list of stuff to talk about tomorrow.