More Testing!
HandleMovement Commands is now separate from the other commands, because
of how many tests it is.

Other Actions were tested more as well.

Game and MapTile were changed to get closer to working with the tests.
I'm not 100% sure what's still wrong with some of them, but others make
sense that they're failing.