More JUnit, More Game (Much Wow)
I edited Game to work better with the JUnit tests, especially for
movement. Also in order to use the check_map() and such methods, and a
method to check if in the correct mode for commands.

There are still tests failing that I don't think should be...

Anyway, MapTile also got some changes since it's involved in the
movement commands, which is where my focus has been.