Game Changer
- In Game, I changed the message for an invalid command because it
sounded more like the command could be used elsewhere.
- GameHandleCommandTest now tests for null commands and commands with
too many arguments (that they respond with the proper messages).

It's currently setup that the stuff is checked to be in the dialog, even
though a message is returned by handle_command. This is because I
thought we agreed on just using dialog and not having a special
errorMessage (or other extra messages). Will bring up as soon as Matt
comes back (in case anyone's reading this today).